+++ beta/app.json
@@ -1,6 +1,6 @@
 {
   "expo": {
-    "name": "hex-works-react",
+    "name": "Hex Works",
     "slug": "hex-works-react",
     "version": "1.0.0",
     "orientation": "portrait",
