+++ beta/src/components/script-panel/providers/script-provider.ts
